+++ packages/builders/radix-icons-builder/build.ts
@@ -1,6 +1,6 @@
-import { createIcons } from '@steeze-ui/icons';
+import { createLibIcons } from '@steeze-ui/icons'
 
-const inputIconThemes = './themes';
-const outputIcons = 'icons';
+const inputIconThemes = './themes'
+const outputIcons = 'icons'
 
-createIcons(inputIconThemes, outputIcons);
+createLibIcons(inputIconThemes, outputIcons)