+++ packages/core/src/builder/index.ts
@@ -5,18 +5,28 @@ import {
 	readFileSync,
 	existsSync,
 	rmSync,
-	mkdirSync
+	mkdirSync,
+	writeFileSync
 } from 'fs'
 import { join } from 'path'
 import { EOL } from 'os'
 import pascalcase from 'pascalcase'
 import { parse } from 'html-parse-stringify'
 
-let svgDict = {}
+let svgDict: {
+	[iconName: string]: {
+		[theme: string]: any
+	}
+} = {}
 const libDir = './src/lib'
 const exportsFile = 'index.js'
 
-export function createIcons(inputIconThemes, outputIcons) {
+interface Options {
+	extendSvgAttributes?: { [key: string]: string }
+	excludeSvgAttributes?: string[]
+}
+
+export function createLibIcons(inputIconThemes: string, outputIcons: string, options: Options) {
 	// check if input dir exists
 	if (!existsSync(inputIconThemes)) {
 		console.log('No themes directory found')
@@ -41,7 +51,7 @@ export function createIcons(inputIconThemes, outputIcons) {
 
 	//collect svg in dict
 	readdirSync(inputIconThemes).forEach((theme) => {
-		getIconsFromTheme(inputIconThemes, theme)
+		getIconsFromTheme(inputIconThemes, theme, options)
 	})
 
 	//write svg dict to files
@@ -51,12 +61,8 @@ export function createIcons(inputIconThemes, outputIcons) {
 	writeExportsModule(outputIcons)
 }
 
-export function getIconsFromTheme(outputExports: string, theme: string) {
-	let allChildTags = 0
-	let usedChildTags = 0
-	let usedPathAttrs = []
-	let usedRectAttrs = []
-	let usedCircleAttrs = []
+export function getIconsFromTheme(outputExports: string, theme: string, options: Options) {
+	let usedAttrsDict: { [key: string]: string[] } = {}
 	readdirSync(join(outputExports, theme)).forEach((fileName) => {
 		const key = pascalcase(fileName.replace('.svg', ''))
 		const data = readFileSync(join(outputExports, theme, fileName)).toString()
@@ -68,49 +74,76 @@ export function getIconsFromTheme(outputExports: string, theme: string) {
 
 		svgDict[key][theme] = { a: svgAst.attrs }
 
-		const childTags = svgAst.children.filter((e) => {
+		if (options.excludeSvgAttributes) {
+			options.excludeSvgAttributes.forEach((attr) => {
+				delete svgDict[key][theme]['a'][attr]
+			})
+		}
+		if (options.extendSvgAttributes) {
+			svgDict[key][theme]['a'] = Object.assign(
+				svgDict[key][theme]['a'],
+				options.extendSvgAttributes
+			)
+		}
+
+		const childTags = svgAst.children.filter((e: any) => {
 			return e.type != 'text'
 		})
-		allChildTags += childTags.length
 
-		const pathAttrs = childTags.filter((e) => e.name === 'path').map((e) => e.attrs)
-		usedChildTags += pathAttrs.length
-		pathAttrs.forEach((path) => {
-			for (let key of Object.keys(path)) {
-				if (!usedPathAttrs.includes(key)) {
-					usedPathAttrs.push(key)
+		let tagsDict: { [key: string]: { [key: string]: string }[] } = {}
+
+		childTags.forEach(
+			({
+				name,
+				attrs,
+				children
+			}: {
+				name: string
+				attrs: { [key: string]: string }
+				children: { name: string; attrs: { [key: string]: string } }[]
+			}) => {
+				if (name === 'g') {
+					if (children) {
+						children.forEach((child) => {
+							child.attrs = Object.assign(attrs, child.attrs)
+
+							if (!tagsDict[child.name]) {
+								tagsDict[child.name] = []
+							}
+							tagsDict[child.name].push(child.attrs)
+						})
+					}
+				} else {
+					if (!tagsDict[name]) {
+						tagsDict[name] = []
+					}
+					tagsDict[name].push(attrs)
 				}
 			}
-		})
+		)
 
-		const rectAttrs = childTags.filter((e) => e.name === 'rect').map((e) => e.attrs)
-		usedChildTags += rectAttrs.length
-		rectAttrs.forEach((rect) => {
-			for (let key of Object.keys(rect)) {
-				if (!usedRectAttrs.includes(key)) {
-					usedRectAttrs.push(key)
-				}
+		svgDict[key][theme] = Object.assign(svgDict[key][theme], tagsDict)
+
+		Object.keys(tagsDict).forEach((tagName) => {
+			if (!usedAttrsDict[tagName]) {
+				usedAttrsDict[tagName] = []
 			}
+			tagsDict[tagName].forEach((attrs) => {
+				Object.keys(attrs).forEach((attr) => {
+					if (!usedAttrsDict[tagName].includes(attr)) {
+						usedAttrsDict[tagName].push(attr)
+					}
+				})
+			})
 		})
-
-		const circleAttrs = childTags.filter((e) => e.name === 'circle').map((e) => e.attrs)
-		usedChildTags += circleAttrs.length
-		circleAttrs.forEach((circle) => {
-			for (let key of Object.keys(circle)) {
-				if (!usedCircleAttrs.includes(key)) {
-					usedCircleAttrs.push(key)
-				}
-			}
-		})
-
-		svgDict[key][theme].path = pathAttrs
-		svgDict[key][theme].rect = rectAttrs
-		svgDict[key][theme].circle = circleAttrs
 	})
-	console.log(`${theme}: ${usedChildTags}/${allChildTags} child tags collected`)
-	console.log('usedPathAttrs', usedPathAttrs)
-	console.log('usedRectAttrs', usedRectAttrs)
-	console.log('usedCircleAttrs', usedCircleAttrs)
+
+	Object.keys(usedAttrsDict).forEach((tagName) => {
+		console.log('theme:', theme)
+
+		console.log(`   <${tagName}>`, usedAttrsDict[tagName])
+	})
+	console.log('')
 }
 
 export async function writeSvgDict(outputIcons: string) {
@@ -118,7 +151,7 @@ export async function writeSvgDict(outputIcons: string) {
 		writeFile(
 			join(join(libDir, outputIcons), `${name}.js`),
 			'export default ' + JSON.stringify(svgDict[name]),
-			(err: any) => {}
+			(_err: any) => {}
 		)
 	})
 }
@@ -135,4 +168,39 @@ export function writeExportsModule(outputIcons: string) {
 	logger.write(exports)
 	logger.write(EOL)
 	logger.end()
+}
+
+export function createThemesFromSuffix(
+	inputDir: string,
+	outputDir: string,
+	themesMap: { [key: string]: string }
+) {
+	if (!existsSync(join(inputDir))) {
+		console.log('No input directory found')
+		return
+	}
+	if (existsSync(join(outputDir))) {
+		rmSync(join(outputDir), { recursive: true })
+	}
+	mkdirSync(join(outputDir))
+
+	const unrecognizedFiles: string[] = []
+	readdirSync(join(inputDir)).forEach((fileName) => {
+		const detectedTheme = Object.keys(themesMap).find((key) => fileName.includes(key))
+		if (detectedTheme) {
+			const data = readFileSync(join(inputDir, fileName)).toString()
+			if (!existsSync(join(outputDir, themesMap[detectedTheme]))) {
+				mkdirSync(join(outputDir, themesMap[detectedTheme]))
+			}
+			writeFileSync(
+				join(outputDir, themesMap[detectedTheme], fileName.replace(detectedTheme, '')),
+				data
+			)
+		} else {
+			unrecognizedFiles.push(fileName)
+		}
+	})
+	if (unrecognizedFiles.length > 0) {
+		console.log('Unrecognized files:', unrecognizedFiles.join(', '))
+	}
 }